Enhance states field replacement function
Keep custom classes and data-* attributes.

When the "States" field is replaced all custom classes and data-*
attributes are lost.

It's a problem for custom styling and js interaction.

For instance if I use a class to size my input, it is removed and the
sizing is not respected anymore.