Modified so it would select the first available option if some are out of stock
Whenever a product with option had an option out of stock, the "out of
stock" label would appear even if the option selected on load was not
out of stock. It gave the impression the product was out of stock.
This modification deals with that.
If there is only one option, choices that are out of stock are disabled
in the select list and "out of stock" is added next to their label. The
browser automatically selects the first available option.
If there are more than 1 option, the code looks for the first available
variation and selects it across all options choices.