Add configuration to allow maximum variations to be adjusted above 50.
Move max exceeded warning to the top of the variations section so it's more likely to be seen. Switched to danger warning since this potentially looses data when saved.

Revise dashboard product page to use JSON for product variations, thus reducing the number of input variables to avoid exceeding max_input_vars.

Set max value on variation form fields to prevent database exception when the values entered overflow the size that the field can cope with.