fix for first checkbox in multicheckbox set
for some reason if the first checkbox was selected its value would be
added to the attribute array twice which is POSTed to the server. This would
then result in a 500 error due to a duplicate entry in the atSelectOptionsSelected
table. This is a quick workaround which checks if the field has already been added
to the DOM